# cm.klg.service-provider/specs/openapi/inbound/main.yml
---
openapi: "3.0.3"
info:
  title: "SERVICE PROVIDER MANAGEMENT REST API"
  description: "API for managing service provider"
  version: "0.0.1-SNAPSHOT"
paths:
  /service-provider:
    post:
      tags:
      - "ServiceProvider"
      summary: "Become Service Provider"
      description: "Become Service Provider"
      operationId: "becomeServiceProvider"
      security:
      - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceProviderRegister"
      responses:
        "201":
          description: "Successful operation"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreationResponse"
        "404":
          description: "SERVICE_PROVIDER_404_001 | SERVICE-PROVIDER_404_003"
        "409":
          description: "SERVICE-PROVIDER_409_001| SERVICE-PROVIDER_409_002"
    get:
      tags:
      - "ServiceProvider"
      summary: "Get All Service Provider"
      description: "Get All Service Provider"
      operationId: "getAllServiceProvider"
      security:
      - oAuth:
        - "service-provider:read:all"
      parameters:
      - name: "status"
        in: "query"
        required: false
        schema:
          $ref: "#/components/schemas/ServiceProviderStatus"
      - name: "page"
        in: "query"
        required: false
        schema:
          type: "integer"
          minimum: "0"
      - name: "limit"
        in: "query"
        required: true
        description: "represents the number of service provider to fetch per page."
        schema:
          type: "integer"
          minimum: "10"
      responses:
        "200":
          description: "Successful operation"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceProviderPaginate"
  /service-provider/search:
    get:
      tags:
      - "ServiceProvider"
      summary: "Search providers with location priority"
      description: "Search providers with location priority (Quarter > District >\
        \ City)"
      operationId: "searchServiceProviders"
      parameters:
      - name: "serviceTypeId"
        in: "query"
        required: true
        schema:
          type: "string"
          format: "uuid"
      - name: "cityId"
        in: "query"
        required: true
        schema:
          type: "string"
          format: "uuid"
      - name: "districtId"
        in: "query"
        required: false
        schema:
          type: "string"
          format: "uuid"
      - name: "quarterId"
        in: "query"
        required: false
        schema:
          type: "string"
          format: "uuid"
      - name: "page"
        in: "query"
        required: false
        schema:
          type: "integer"
          minimum: "0"
      - name: "limit"
        in: "query"
        required: true
        schema:
          type: "integer"
          minimum: "10"
      responses:
        "200":
          description: "Successful operation"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceProviderPaginate"
  /service-provider/{serviceProviderId}:
    get:
      tags:
      - "ServiceProvider"
      summary: "Get Service Provider By Id"
      description: "Get Service Provider By Id"
      operationId: "GetServiceProviderById"
      parameters:
      - name: "serviceProviderId"
        in: "path"
        required: true
        schema:
          type: "string"
          format: "uuid"
      - name: "userId"
        in: "query"
        required: false
        schema:
          type: "string"
          format: "uuid"
      responses:
        "200":
          description: "Successful operation"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceProviderPublicProfile"
        "404":
          description: "SERVICE_PROVIDER_404_002"
  /service-provider/{serviceProviderId}/approve:
    put:
      tags:
      - "ServiceProvider"
      summary: "Approve Service Provider Request"
      description: "Approve Service Provider Request"
      operationId: "ApproveServiceProvider"
      security:
      - oAuth:
        - "service-provider:approve"
      parameters:
      - name: "serviceProviderId"
        in: "path"
        required: true
        schema:
          type: "string"
          format: "uuid"
      responses:
        "204":
          description: "Successful operation"
        "404":
          description: "SERVICE_PROVIDER_404_002"
  /service-provider/{serviceProviderId}/reject:
    put:
      tags:
      - "ServiceProvider"
      summary: "Reject Service Provider Request"
      description: "Reject Service Provider Request"
      operationId: "RejectServiceProvider"
      security:
      - oAuth:
        - "service-provider:reject"
      parameters:
      - name: "serviceProviderId"
        in: "path"
        required: true
        schema:
          type: "string"
          format: "uuid"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectionReason"
      responses:
        "204":
          description: "Successful operation"
        "404":
          description: "SERVICE_PROVIDER_404_002"
  /service-provider/profile:
    get:
      tags:
      - "ServiceProvider"
      summary: "Get Service Provider Profile"
      description: "Get Service Provider Profile"
      operationId: "getServiceProviderProfile"
      security:
      - bearerAuth: []
      responses:
        "200":
          description: "Successful operation"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceProvider"
        "404":
          description: "SERVICE_PROVIDER_404_002"
    put:
      tags:
      - "ServiceProvider"
      summary: "Update Service Provider Profile"
      description: "Update the authenticated service provider's contact details, location\
        \ and description."
      operationId: "updateServiceProviderProfile"
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateServiceProviderProfileRequest"
      responses:
        "204":
          description: "Successful operation"
        "404":
          description: "SERVICE_PROVIDER_404_002"
  /service-provider/portfolio:
    post:
      tags:
      - "ServiceProvider"
      summary: "Add Provider's Portfolio"
      description: "Add Provider's Portfolio"
      operationId: "addPortfolioItem"
      security:
      - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePortfolioItemRequest"
      responses:
        "201":
          description: "Successful operation"
        "404":
          description: "SERVICE_PROVIDER_404_002"
    get:
      tags:
      - "ServiceProvider"
      summary: "Get All Provider's Portfolio"
      description: "Get All Provider's Portfolio"
      operationId: "getAllMyPortfolioItem"
      security:
      - bearerAuth: []
      responses:
        "200":
          description: "Successful operation"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/PortfolioItem"
  /service-provider/{serviceProviderId}/portfolio:
    get:
      tags:
      - "ServiceProvider"
      summary: "Get provider portfolio"
      description: "Retrieve all portfolio items of the specified service provider."
      operationId: "getProviderPortfolio"
      parameters:
      - name: "serviceProviderId"
        in: "path"
        required: true
        schema:
          type: "string"
          format: "uuid"
      responses:
        "200":
          description: "Successful operation"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/PortfolioItem"
  /service-provider/portfolio/{portfolioId}:
    put:
      tags:
      - "ServiceProvider"
      summary: "Update a portfolio item"
      operationId: "updatePortfolioItem"
      security:
      - bearerAuth: []
      parameters:
      - name: "portfolioId"
        in: "path"
        required: true
        schema:
          type: "string"
          format: "uuid"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePortfolioItemRequest"
      responses:
        "204":
          description: "Successful operation"
        "404":
          description: "SERVICE_PROVIDER_404_002"
    delete:
      tags:
      - "ServiceProvider"
      summary: "Delete a portfolio item"
      operationId: "deletePortfolioItem"
      security:
      - bearerAuth: []
      parameters:
      - name: "portfolioId"
        in: "path"
        required: true
        schema:
          type: "string"
          format: "uuid"
      responses:
        "204":
          description: "Successful operation"
        "404":
          description: "SERVICE_PROVIDER_404_002"
  /service-provider/services:
    get:
      tags:
      - "ServiceProvider"
      summary: "Get my services"
      operationId: "getMyServices"
      security:
      - bearerAuth: []
      responses:
        "200":
          description: "Successful operation"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/UserService"
    post:
      tags:
      - "ServiceProvider"
      summary: "Add a service to provide"
      operationId: "addNewService"
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceType"
      responses:
        "201":
          description: "Successful operation"
        "404":
          description: "SERVICE_PROVIDER_404_002 || SERVICE-PROVIDER_404_003"
  /service-provider/{serviceProviderId}/services:
    get:
      tags:
      - "ServiceProvider"
      summary: "Get a service provider's services"
      operationId: "getProviderServices"
      parameters:
      - name: "serviceProviderId"
        in: "path"
        required: true
        schema:
          type: "string"
          format: "uuid"
      responses:
        "200":
          description: "Successful operation"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/UserService"
        "404":
          description: "SERVICE_PROVIDER_404_002"
  /service-catalog:
    get:
      tags:
      - "ServiceType"
      summary: "Get the complete service catalog grouped by categories"
      description: "Get the complete service catalog grouped by categories"
      operationId: "getServiceCatalog"
      responses:
        "200":
          description: "Successful operation"
          content:
            application/json:
              schema:
                type: "object"
                additionalProperties:
                  type: "array"
                  items:
                    $ref: "#/components/schemas/ServiceCatalogItem"
components:
  schemas:
    UpdateServiceProviderProfileRequest:
      required:
      - "phoneNumber"
      - "city"
      - "district"
      - "quarter"
      type: "object"
      properties:
        city:
          type: "string"
          format: "uuid"
        district:
          type: "string"
          format: "uuid"
        quarter:
          type: "string"
          format: "uuid"
        about:
          type: "string"
        phoneNumber:
          $ref: "#/components/schemas/PhoneNumber"
    RejectionReason:
      type: "object"
      properties:
        reason:
          type: "string"
    ServiceProviderRegister:
      required:
      - "phoneNumber"
      - "city"
      - "district"
      - "quarter"
      - "serviceType"
      type: "object"
      properties:
        city:
          type: "string"
          format: "uuid"
        district:
          type: "string"
          format: "uuid"
        quarter:
          type: "string"
          format: "uuid"
        about:
          type: "string"
        phoneNumber:
          $ref: "#/components/schemas/PhoneNumber"
        serviceType:
          $ref: "#/components/schemas/ServiceType"
    ServiceProviderPaginate:
      type: "object"
      required:
      - "serviceProvider"
      - "count"
      properties:
        serviceProvider:
          type: "array"
          items:
            $ref: "#/components/schemas/ServiceProvider"
        count:
          type: "integer"
          format: "int64"
    ServiceProviderPublicProfile:
      required:
      - "id"
      - "userId"
      - "lastname"
      - "city"
      - "district"
      - "quarter"
      - "userIsProviderClient"
      - "serviceProviderStatus"
      - "createdAt"
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
        userId:
          type: "string"
          format: "uuid"
        firstname:
          type: "string"
        lastname:
          type: "string"
        city:
          type: "string"
          format: "uuid"
        district:
          type: "string"
          format: "uuid"
        quarter:
          type: "string"
          format: "uuid"
        about:
          type: "string"
        createdAt:
          type: "string"
          format: "date-time"
        updatedAt:
          type: "string"
          format: "date-time"
        phoneNumber:
          $ref: "#/components/schemas/PhoneNumber"
        userServices:
          type: "array"
          items:
            $ref: "#/components/schemas/UserService"
        portfolio:
          type: "array"
          items:
            $ref: "#/components/schemas/PortfolioItem"
    ServiceProvider:
      required:
      - "id"
      - "userId"
      - "lastname"
      - "city"
      - "district"
      - "quarter"
      - "phoneNumber"
      - "createdAt"
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
        userId:
          type: "string"
          format: "uuid"
        firstname:
          type: "string"
        lastname:
          type: "string"
        city:
          type: "string"
          format: "uuid"
        district:
          type: "string"
          format: "uuid"
        quarter:
          type: "string"
          format: "uuid"
        about:
          type: "string"
        createdAt:
          type: "string"
          format: "date-time"
        updatedAt:
          type: "string"
          format: "date-time"
        phoneNumber:
          $ref: "#/components/schemas/PhoneNumber"
    UserService:
      type: "object"
      required:
      - "serviceName"
      - "serviceCategory"
      - "yearOfExperience"
      - "document"
      - "createdAt"
      properties:
        serviceName:
          type: "string"
        serviceCategory:
          type: "string"
        yearOfExperience:
          type: "integer"
        document:
          type: "string"
          format: "uuid"
        createdAt:
          type: "string"
          format: "date-time"
        serviceType:
          $ref: "#/components/schemas/ServiceType"
    PhoneNumber:
      type: "object"
      required:
      - "countryCode"
      - "number"
      properties:
        countryCode:
          type: "string"
        number:
          type: "string"
    ServiceType:
      type: "object"
      required:
      - "id"
      - "yearOfExperience"
      - "document"
      properties:
        id:
          type: "string"
          format: "uuid"
        yearOfExperience:
          type: "integer"
        document:
          type: "string"
          format: "uuid"
    ServiceProviderStatus:
      type: "string"
      enum:
      - "PENDING"
      - "APPROVED"
      - "REJECTED"
    ServiceCatalogItem:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
        name:
          type: "string"
        category:
          type: "string"
    CreatePortfolioItemRequest:
      type: "object"
      required:
      - "title"
      - "description"
      - "mediaId"
      properties:
        title:
          type: "string"
          minLength: 3
          maxLength: 100
        description:
          type: "string"
          minLength: 10
          maxLength: 500
        mediaId:
          type: "string"
          format: "uuid"
    PortfolioItem:
      type: "object"
      required:
      - "id"
      - "serviceProviderId"
      - "title"
      - "description"
      - "mediaId"
      - "createdAt"
      properties:
        id:
          type: "string"
          format: "uuid"
        title:
          type: "string"
        description:
          type: "string"
        mediaId:
          type: "string"
          format: "uuid"
        createdAt:
          type: "string"
          format: "date-time"
    CreationResponse:
      type: "object"
      required:
      - "newId"
      properties:
        newId:
          type: "string"
          format: "uuid"